==> models/saiva-3-day-hosp-notes-experiment/src/notebooks/04-nlp-features.ipynb <==
import sys
import pandas as pd
from pathlib import Path
sys.path.insert(0, '/src')
from shared.generate_nlp_features import NERFeature, TopicModelFeature, preprocess_nlp_data
from eliot import start_action, start_task, to_file, log_message
from shared.utils import get_client_class, get_memory_usage
from shared.constants import CLIENT
to_file(sys.stdout)

processed_path = Path('/data/processed')
raw_path = Path('/data/raw')

notes_df = pd.read_parquet(raw_path/'patient_progress_notes.parquet')
base_df = pd.read_parquet(processed_path/'03-result.parquet')

print(notes_df.shape)
print(base_df.shape)

notes_df = preprocess_nlp_data(notes_df)
notes_df.to_parquet(processed_path/'04-preprocessed_notes.parquet')
print(notes_df.shape)

# notes_df = pd.read_parquet(processed_path/'04-preprocessed_notes.parquet')

%%time

ner_obj = NERFeature(notes_df.copy())
base_df = ner_obj.execute(base_df)

base_df.to_parquet(processed_path/'04a-result.parquet')

### Avg Run time : 28min 13s

# base_df = pd.read_parquet(processed_path/'04a-result.parquet')

%%time

topic_obj = TopicModelFeature(notes_df=notes_df, name=CLIENT)
base_df = topic_obj.execute(base_df)

print(get_memory_usage(base_df))

base_df.to_parquet(processed_path/'04-result.parquet')
